Rename the AtUproot file to differ from the class
