+++ atuproot/__init__.py
@@ -1,3 +1,3 @@
-from .AtUproot import AtUproot
+from .atuproot_main import AtUproot
 from .EventBuilder import EventBuilder
 from .EventBuilderConfigMaker import EventBuilderConfigMaker
